requirements.txt
biopython
psycopg2-binary
python-dotenv